requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
urllib3==1.26.15
yarl==1.9.2
